share logic for handle creation
